feat: move to server components
