+++ apps/nextjs/src/utils/index.ts
@@ -1,0 +1,82 @@
+export function areSameDates(date1: Date, date2: Date) {
+  return (
+    date1.getFullYear() === date2.getFullYear() &&
+    date1.getMonth() === date2.getMonth() &&
+    date1.getDate() === date2.getDate()
+  );
+}
+
+export function getShortDay(date: Date) {
+  const formatter = new Intl.DateTimeFormat("en-US", {
+    weekday: "short",
+  });
+  return formatter.format(date);
+}
+
+export function getDateSlug(date: Date) {
+  return `/${new Intl.DateTimeFormat("en-US").format(date)}`;
+}
+
+export function getPrevDay(date: Date) {
+  const prevDay = new Date(date);
+  prevDay.setDate(date.getDate() - 1);
+  return prevDay;
+}
+
+export function getNextDay(date: Date) {
+  const nextDay = new Date(date);
+  nextDay.setDate(date.getDate() + 1);
+  return nextDay;
+}
+
+export const converMinutesToHours = (totalMinutes: number) => {
+  const hours = Math.floor(totalMinutes / 60);
+  const minutes = totalMinutes % 60;
+
+  return `${padToTwoDigits(hours)}:${padToTwoDigits(minutes)}`;
+};
+
+export const currentWeekBoundaries = getWeekBoundaries(new Date());
+
+function padToTwoDigits(num: number) {
+  return num.toString().padStart(2, "0");
+}
+export const currentWeekDates = getWeekDates(
+  currentWeekBoundaries.startDate,
+  currentWeekBoundaries.endDate,
+);
+
+function getWeekDates(startDate: Date, endDate: Date) {
+  let dates = [];
+  let currentDate = new Date(startDate);
+
+  while (currentDate <= endDate) {
+    dates.push(new Date(currentDate));
+    currentDate.setDate(currentDate.getDate() + 1);
+  }
+
+  return dates;
+}
+
+function getWeekBoundaries(date: Date) {
+  // Calculate start date of the week (Monday)
+  let startDate = new Date(date);
+  startDate.setDate(date.getDate() - ((date.getDay() + 6) % 7));
+
+  // Set hours, minutes, seconds, and milliseconds to 0
+  startDate.setHours(0);
+  startDate.setMinutes(0);
+  startDate.setSeconds(0);
+  startDate.setMilliseconds(0);
+
+  // Calculate end date of the week (Sunday)
+  let endDate = new Date(startDate);
+  endDate.setDate(startDate.getDate() + 6);
+
+  // Set hours to 23, minutes to 59, seconds to 59
+  endDate.setHours(23);
+  endDate.setMinutes(59);
+  endDate.setSeconds(59);
+
+  return { startDate, endDate };
+}
